Refactor CreateRenter page
