+++ .eslintrc.json
@@ -42,7 +42,8 @@
           ]
         }
       ],
-      "react/jsx-one-expression-per-line": "off"
+      "react/jsx-one-expression-per-line": "off",
+      "react/prop-types": "off"
     },
     "settings": {
         "react": {
