+++ src/common/reportError.ts
@@ -1,4 +1,4 @@
-export default function reportError(e: string): void {
+export default function reportError(e: Error): void {
   // eslint-disable-next-line no-console
   console.error(e);
 }
