+++ src/components/NewRenter/CreateRenter.tsx
@@ -11,139 +11,42 @@ import {
   InputLabel,
   FormControl,
   Button,
-  Grid,
-  Typography,
   Snackbar,
 } from '@material-ui/core';
-import Autocomplete from '@material-ui/lab/Autocomplete';
-import Alert, { Color } from '@material-ui/lab/Alert';
-import { makeStyles, Theme, createStyles } from '@material-ui/core/styles';
+import { AutocompleteRenderInputParams } from '@material-ui/lab/Autocomplete';
+import Alert from '@material-ui/lab/Alert';
 import { useIntl, FormattedMessage } from 'react-intl';
 import phone from 'phone';
-import parse from 'autosuggest-highlight/parse';
-import throttle from 'lodash/throttle';
 
 // Locale
 import messages from './messages';
 
 // Internal
 import { postTenants } from '~/api/coreApi';
-import config from '~/config';
+import GoogleAutocomplete from './GoogleAutocomplete';
 
 import Renter = Models.Renter;
-
-function loadScript(src: string, position: HTMLElement | null, id: string): void {
-  if (!position) {
-    return;
-  }
-
-  const script = document.createElement('script');
-  script.setAttribute('async', '');
-  script.setAttribute('id', id);
-  script.src = src;
-  position.appendChild(script);
-}
-
-const useStyles = makeStyles((theme: Theme) => createStyles({
-  formControl: {
-    minWidth: 100,
-  },
-  submitButton: {
-    marginLeft: theme.spacing(1),
-  },
-}));
-
-const autocompleteService = { current: null };
+import PlaceType = Models.PlaceType;
 
 const defaultValidationForm = {
   phone: '',
 };
 
-interface PlaceType {
-  description: string;
-  place_id: string;
-  structured_formatting: {
-    main_text: string;
-    secondary_text: string;
-    main_text_matched_substrings: [
-      {
-        offset: number;
-        length: number;
-      },
-    ];
-  };
-}
+const CURRENCY = {
+  USD: 'USD',
+  EUR: 'EUR',
+  UAH: 'UAH',
+  RUB: 'RUB',
+};
 
 const CreateRenter: React.FunctionComponent = () => {
   const intl = useIntl();
-  const classes = useStyles();
   const [validationForm, setValidationForm] = useState(defaultValidationForm);
   const [locationValue, setLocationValue] = React.useState<PlaceType | null>(null);
-  const [inputLocationValue, setInputLocationValue] = React.useState('');
-  const [locationOptions, setLocationOptions] = React.useState<PlaceType[]>([]);
   const [isSendingForm, setIsSendingForm] = React.useState(false);
-  const [openAlert, setOpenAlert] = React.useState<Color>();
-  const loaded = React.useRef(false);
+  const [openAlert, setOpenAlert] = React.useState(false);
 
   const history = useHistory();
-
-  if (typeof window !== 'undefined' && !loaded.current) {
-    if (!document.querySelector('#google-maps')) {
-      loadScript(
-        `https://maps.googleapis.com/maps/api/js?key=${config.google.placeApi}&libraries=places`,
-        document.querySelector('head'),
-        'google-maps',
-      );
-    }
-
-    loaded.current = true;
-  }
-
-  const fetch = React.useMemo(
-    () => throttle((
-      request: { input: string; types: string[] },
-      callback: (results?: PlaceType[]) => void,
-    ) => {
-      (autocompleteService.current as any).getPlacePredictions(request, callback);
-    }, 200),
-    [],
-  );
-
-  React.useEffect(() => {
-    let active = true;
-
-    if (!autocompleteService.current && (window as any).google) {
-      autocompleteService.current = new (window as any).google.maps.places.AutocompleteService();
-    }
-    if (!autocompleteService.current) {
-      return undefined;
-    }
-
-    if (inputLocationValue === '') {
-      setLocationOptions(locationValue ? [locationValue] : []);
-      return undefined;
-    }
-
-    fetch({ input: inputLocationValue, types: ['(cities)'] }, (results?: PlaceType[]) => {
-      if (active) {
-        let newOptions = [] as PlaceType[];
-
-        if (locationValue) {
-          newOptions = [locationValue];
-        }
-
-        if (results) {
-          newOptions = [...newOptions, ...results];
-        }
-
-        setLocationOptions(newOptions);
-      }
-    });
-
-    return (): void => {
-      active = false;
-    };
-  }, [locationValue, inputLocationValue, fetch]);
 
   const validateForm = (formData: FormData): boolean => {
     let isValid = true;
@@ -194,86 +97,41 @@ const CreateRenter: React.FunctionComponent = () => {
       if (response) {
         history.push(`/ad/${response.data.id}`);
       } else {
-        setOpenAlert('error');
+        setOpenAlert(true);
         setIsSendingForm(false);
       }
     } catch (e) {
-      setOpenAlert('error');
+      setOpenAlert(true);
       setIsSendingForm(false);
     }
   };
 
   return (
-    <Container maxWidth="sm">
+    <Container maxWidth="sm" id="new-renter-page">
       <form onSubmit={handleSubmit} onChange={(): void => setValidationForm(defaultValidationForm)}>
         <TextField name="title" label={intl.formatMessage(messages.title)} fullWidth required />
         <Box display="flex" justifyContent="space-between">
-          <TextField name="name" label={intl.formatMessage(messages.userName)} required />
+          <TextField name="name" label={intl.formatMessage(messages.userName)} required fullWidth />
           <TextField
             error={!!validationForm.phone}
             name="phone"
             label={intl.formatMessage(messages.userPhone)}
             required
             helperText={validationForm.phone}
+            className="phone-input"
+            fullWidth
           />
-          <TextField name="social" label={intl.formatMessage(messages.social)} />
+          <TextField name="social" label={intl.formatMessage(messages.social)} fullWidth />
         </Box>
         <Box display="flex" justifyContent="space-between">
-          <Autocomplete
-            id="google-map"
-            autoHighlight
-            getOptionLabel={(option): string => (typeof option === 'string' ? option : option.description)}
-            filterOptions={(x): PlaceType[] => x}
-            style={{
-              maxWidth: 409,
-              flex: 1,
-            }}
-            options={locationOptions}
-            autoComplete
-            includeInputInList
-            filterSelectedOptions
-            value={locationValue}
-            onChange={(event: any, newValue: PlaceType | null): void => {
-              setLocationOptions(newValue ? [newValue, ...locationOptions] : locationOptions);
-              setLocationValue(newValue);
-            }}
-            onInputChange={(event, newInputValue): void => {
-              setInputLocationValue(newInputValue);
-            }}
-            renderInput={(params): React.ReactNode => (
+          <GoogleAutocomplete
+            onChange={(value): void => setLocationValue(value)}
+            renderInput={(params: AutocompleteRenderInputParams): React.ReactNode => (
               // eslint-disable-next-line react/jsx-props-no-spreading
               <TextField {...params} label={intl.formatMessage(messages.city)} fullWidth required />
             )}
-            renderOption={(option): React.ReactNode => {
-              const matches = option.structured_formatting.main_text_matched_substrings;
-              const parts = parse(
-                option.structured_formatting.main_text,
-                matches.map((match: any) => [match.offset, match.offset + match.length]),
-              );
-
-              return (
-                <Grid container alignItems="center">
-                  <Grid item xs>
-                    {parts.map((part, index) => (
-                      <span
-                        // eslint-disable-next-line react/no-array-index-key
-                        key={index}
-                        style={{
-                          fontWeight: part.highlight ? 700 : 400,
-                        }}
-                      >
-                        {part.text}
-                      </span>
-                    ))}
-                    <Typography variant="body2" color="textSecondary">
-                      {option.structured_formatting.secondary_text}
-                    </Typography>
-                  </Grid>
-                </Grid>
-              );
-            }}
           />
-          <FormControl className={classes.formControl}>
+          <FormControl className="form-control" fullWidth>
             <InputLabel id="select-housing-type-label"><FormattedMessage {...messages.housingType} /></InputLabel>
             <Select
               name="housingType"
@@ -288,18 +146,23 @@ const CreateRenter: React.FunctionComponent = () => {
           </FormControl>
         </Box>
         <Box display="flex" justifyContent="space-between">
-          <TextField name="minBudget" label={intl.formatMessage(messages.budgetFrom)} type="number" required />
-          <TextField name="maxBudget" label={intl.formatMessage(messages.budgetTo)} type="number" required />
-          <FormControl className={classes.formControl}>
+          <TextField name="minBudget" label={intl.formatMessage(messages.budgetFrom)} type="number" required fullWidth />
+          <TextField
+            name="maxBudget"
+            label={intl.formatMessage(messages.budgetTo)}
+            type="number"
+            required
+            fullWidth
+            className="budget-to-input"
+          />
+          <FormControl className="form-control" fullWidth>
             <InputLabel id="select-currency-label"><FormattedMessage {...messages.currency} /></InputLabel>
             <Select
               name="currency"
               labelId="select-currency-label"
               defaultValue="USD"
             >
-              <MenuItem value="USD">USD</MenuItem>
-              <MenuItem value="UAH">UAH</MenuItem>
-              <MenuItem value="EUR">EUR</MenuItem>
+              {(Object.values(CURRENCY)).map(item => <MenuItem value={item}>{item}</MenuItem>)}
             </Select>
           </FormControl>
         </Box>
@@ -335,7 +198,7 @@ const CreateRenter: React.FunctionComponent = () => {
           </Button>
           <Button
             disabled={isSendingForm}
-            className={classes.submitButton}
+            className="submit-button"
             variant="contained"
             color="primary"
             type="submit"
@@ -345,11 +208,11 @@ const CreateRenter: React.FunctionComponent = () => {
         </Box>
       </form>
       <Snackbar
-        open={!!openAlert}
+        open={openAlert}
         autoHideDuration={6000}
-        onClose={(): void => setOpenAlert(undefined)}
+        onClose={(): void => setOpenAlert(false)}
       >
-        <Alert onClose={(): void => setOpenAlert(undefined)} severity={openAlert}>
+        <Alert onClose={(): void => setOpenAlert(false)} severity="error">
           <FormattedMessage {...messages.error} />
         </Alert>
       </Snackbar>
