+++ src/config.ts
@@ -1,8 +1,4 @@
 export default {
-  api: {
-    url: process.env.REACT_APP_API_URL,
-  },
-  google: {
-    placeApi: process.env.REACT_APP_GOOGLE_PLACES_API,
-  },
+  coreApiUrl: process.env.REACT_APP_API_URL,
+  googleApi: process.env.REACT_APP_GOOGLE_PLACES_API,
 };
