+++ src/models/index.d.ts
@@ -20,4 +20,18 @@ declare namespace Models {
     offset?: number;
     limit?: number;
   }
+  export interface PlaceType {
+    description: string;
+    place_id: string;
+    structured_formatting: {
+      main_text: string;
+      secondary_text: string;
+      main_text_matched_substrings: [
+        {
+          offset: number;
+          length: number;
+        },
+      ];
+    };
+  }
 }
