+++ src/scss/_variables.scss
@@ -1,0 +1,2 @@
+// spacing
+$sp-m: 12px;
