+++ src/scss/style.scss
@@ -1,2 +1,3 @@
 @import 'common';
 @import 'catalog';
+@import 'newRenter';
